Feat - 유효성 검사 (#18)
